rename module entry -> dir_entry
