+++ src/filetypes.rs
@@ -1,4 +1,4 @@
-use crate::entry;
+use crate::dir_entry;
 use crate::filesystem;
 
 /// Whether or not to show
@@ -14,7 +14,7 @@ pub struct FileTypes {
 }
 
 impl FileTypes {
-    pub fn should_ignore(&self, entry: &entry::DirEntry) -> bool {
+    pub fn should_ignore(&self, entry: &dir_entry::DirEntry) -> bool {
         if let Some(ref entry_type) = entry.file_type() {
             (!self.files && entry_type.is_file())
                 || (!self.directories && entry_type.is_dir())
